+++ src/main/java/com/data/filtro/interview/ModelRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ModelRepository extends JpaRepository<MyModel, Integer> {
+public interface ModelRepository extends JpaRepository<MyModel, Long> {
     List<MyModel> findAll();
 }
